+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.7.2](https://github.com/formatjs/formatjs/compare/@formatjs/cli@6.7.1...@formatjs/cli@6.7.2) (2025-06-15)
+
+**Note:** Version bump only for package @formatjs/cli
+
 ## [6.7.1](https://github.com/formatjs/formatjs/compare/@formatjs/cli@6.7.0...@formatjs/cli@6.7.1) (2025-05-05)
 
 ### Bug Fixes